+++ src/index.html
@@ -129,10 +129,12 @@
 </template>
 
 
+<script src="js/main.js" charset="utf-8"></script>
 <script src="js/resizer.js" charset="utf-8"></script>
 <script src="js/upload.js" charset="utf-8"></script>
 <script src="js/check.js" charset="utf-8"></script>
-<script src="pictures.js" charset="utf-8"></script>
+<script src="js/pictures.js" charset="utf-8"></script>
+<script src="js/load.js" charset="utf-8"></script>
 
 </body>
 </html>
