+++ src/js/upload.js
@@ -8,323 +8,322 @@
 
 'use strict';
 
-(function() {
+var Resizer = require('resizer');
 
-  /** @enum {string} */
-  var FileType = {
-    'GIF': '',
-    'JPEG': '',
-    'PNG': '',
-    'SVG+XML': ''
-  };
+/** @enum {string} */
+var FileType = {
+  'GIF': '',
+  'JPEG': '',
+  'PNG': '',
+  'SVG+XML': ''
+};
 
-  /** @enum {number} */
-  var Action = {
-    ERROR: 0,
-    UPLOADING: 1,
-    CUSTOM: 2
-  };
+/** @enum {number} */
+var Action = {
+  ERROR: 0,
+  UPLOADING: 1,
+  CUSTOM: 2
+};
 
-  /**
-   * Регулярное выражение, проверяющее тип загружаемого файла. Составляется
-   * из ключей FileType.
-   * @type {RegExp}
-   */
-  var fileRegExp = new RegExp('^image/(' + Object.keys(FileType).join('|').replace('\+', '\\+') + ')$', 'i');
+/**
+ * Регулярное выражение, проверяющее тип загружаемого файла. Составляется
+ * из ключей FileType.
+ * @type {RegExp}
+ */
+var fileRegExp = new RegExp('^image/(' + Object.keys(FileType).join('|').replace('\+', '\\+') + ')$', 'i');
 
-  /**
-   * @type {Object.<string, string>}
-   */
-  var filterMap;
+/**
+ * @type {Object.<string, string>}
+ */
+var filterMap;
 
-  /**
-   * Объект, который занимается кадрированием изображения.
-   * @type {Resizer}
-   */
-  var currentResizer;
+/**
+ * Объект, который занимается кадрированием изображения.
+ * @type {Resizer}
+ */
+var currentResizer;
 
-  /**
-   * Удаляет текущий объект {@link Resizer}, чтобы создать новый с другим
-   * изображением.
-   */
-  var cleanupResizer = function() {
-    if (currentResizer) {
-      currentResizer.remove();
-      currentResizer = null;
-    }
-  };
+/**
+ * Удаляет текущий объект {@link Resizer}, чтобы создать новый с другим
+ * изображением.
+ */
+var cleanupResizer = function() {
+  if (currentResizer) {
+    currentResizer.remove();
+    currentResizer = null;
+  }
+};
 
-  /**
-   * Ставит одну из трех случайных картинок на фон формы загрузки.
-   */
-  var updateBackground = function() {
-    var images = [
-      'img/logo-background-1.jpg',
-      'img/logo-background-2.jpg',
-      'img/logo-background-3.jpg'
-    ];
+/**
+ * Ставит одну из трех случайных картинок на фон формы загрузки.
+ */
+var updateBackground = function() {
+  var images = [
+    'img/logo-background-1.jpg',
+    'img/logo-background-2.jpg',
+    'img/logo-background-3.jpg'
+  ];
 
-    var backgroundElement = document.querySelector('.upload');
-    var randomImageNumber = Math.round(Math.random() * (images.length - 1));
-    backgroundElement.style.backgroundImage = 'url(' + images[randomImageNumber] + ')';
-  };
+  var backgroundElement = document.querySelector('.upload');
+  var randomImageNumber = Math.round(Math.random() * (images.length - 1));
+  backgroundElement.style.backgroundImage = 'url(' + images[randomImageNumber] + ')';
+};
 
-  /**
-   * Кнопка "Продолжить".
-   */
-  var resizeForwardBtn = document.getElementById('resize-fwd');
+/**
+ * Кнопка "Продолжить".
+ */
+var resizeForwardBtn = document.getElementById('resize-fwd');
 
-  /*
-  * Обновляет атрибут disabled у кнопки отправки формы в зависимости от её валидности.
-  */
-  function toggleResizeFormSubmit() {
-    resizeForwardBtn.disabled = !resizeFormIsValid();
+/*
+* Обновляет атрибут disabled у кнопки отправки формы в зависимости от её валидности.
+*/
+function toggleResizeFormSubmit() {
+  resizeForwardBtn.disabled = !resizeFormIsValid();
+}
+
+
+
+/**
+ * Проверяет, валидны ли данные, в форме кадрирования.
+ * @return {boolean}
+ */
+var resizeFormIsValid = function() {
+  var imWidth = currentResizer._image.naturalWidth;
+  var imHeight = currentResizer._image.naturalHeight;
+  var resizeX = 1 * resizeInputX.value;
+  var resizeY = 1 * resizeInputY.value;
+  var resizeSize = 1 * resizeInputSize.value;
+  var isValid = (resizeX + resizeSize <= imWidth) &&
+    (resizeY + resizeSize <= imHeight) && (resizeX >= 0) && (resizeY >= 0);
+
+  resizeForwardBtn.disabled = !isValid;
+
+  return isValid;
+};
+
+/**
+ * Форма загрузки изображения.
+ * @type {HTMLFormElement}
+ */
+var uploadForm = document.forms['upload-select-image'];
+
+/**
+ * Форма кадрирования изображения.
+ * @type {HTMLFormElement}
+ */
+var resizeForm = document.forms['upload-resize'];
+
+/**
+ * Форма добавления фильтра.
+ * @type {HTMLFormElement}
+ */
+var filterForm = document.forms['upload-filter'];
+
+/**
+ * @type {HTMLImageElement}
+ */
+var filterImage = filterForm.querySelector('.filter-image-preview');
+
+/**
+ * @type {HTMLElement}
+ */
+var uploadMessage = document.querySelector('.upload-message');
+
+/**
+ * Объявить поля ввода формы.
+ */
+var resizeInputX = resizeForm.elements.x;
+var resizeInputY = resizeForm.elements.y;
+var resizeInputSize = resizeForm.elements.size;
+
+resizeInputX.oninput = toggleResizeFormSubmit;
+resizeInputY.oninput = toggleResizeFormSubmit;
+resizeInputSize.oninput = toggleResizeFormSubmit;
+
+/**
+ * @param {Action} action
+ * @param {string=} message
+ * @return {Element}
+ */
+var showMessage = function(action, message) {
+  var isError = false;
+
+  switch (action) {
+    case Action.UPLOADING:
+      message = message || 'Кексограмим&hellip;';
+      break;
+
+    case Action.ERROR:
+      isError = true;
+      message = message || 'Неподдерживаемый формат файла<br> <a href="' + document.location + '">Попробовать еще раз</a>.';
+      break;
   }
 
+  uploadMessage.querySelector('.upload-message-container').innerHTML = message;
+  uploadMessage.classList.remove('invisible');
+  uploadMessage.classList.toggle('upload-message-error', isError);
+  return uploadMessage;
+};
 
+var hideMessage = function() {
+  uploadMessage.classList.add('invisible');
+};
 
-  /**
-   * Проверяет, валидны ли данные, в форме кадрирования.
-   * @return {boolean}
-   */
-  var resizeFormIsValid = function() {
-    var imWidth = currentResizer._image.naturalWidth;
-    var imHeight = currentResizer._image.naturalHeight;
-    var resizeX = 1 * resizeInputX.value;
-    var resizeY = 1 * resizeInputY.value;
-    var resizeSize = 1 * resizeInputSize.value;
-    var isValid = (resizeX + resizeSize <= imWidth) &&
-      (resizeY + resizeSize <= imHeight) && (resizeX >= 0) && (resizeY >= 0);
+/**
+ * Обработчик изменения изображения в форме загрузки. Если загруженный
+ * файл является изображением, считывается исходник картинки, создается
+ * Resizer с загруженной картинкой, добавляется в форму кадрирования
+ * и показывается форма кадрирования.
+ * @param {Event} evt
+ */
+uploadForm.onchange = function(evt) {
+  var element = evt.target;
+  if (element.id === 'upload-file') {
+    // Проверка типа загружаемого файла, тип должен быть изображением
+    // одного из форматов: JPEG, PNG, GIF или SVG.
+    if (fileRegExp.test(element.files[0].type)) {
+      var fileReader = new FileReader();
 
-    resizeForwardBtn.disabled = !isValid;
+      showMessage(Action.UPLOADING);
 
-    return isValid;
-  };
+      fileReader.onload = function() {
+        cleanupResizer();
 
-  /**
-   * Форма загрузки изображения.
-   * @type {HTMLFormElement}
-   */
-  var uploadForm = document.forms['upload-select-image'];
+        currentResizer = new Resizer(fileReader.result);
+        currentResizer.setElement(resizeForm);
+        uploadMessage.classList.add('invisible');
 
-  /**
-   * Форма кадрирования изображения.
-   * @type {HTMLFormElement}
-   */
-  var resizeForm = document.forms['upload-resize'];
+        uploadForm.classList.add('invisible');
+        resizeForm.classList.remove('invisible');
 
-  /**
-   * Форма добавления фильтра.
-   * @type {HTMLFormElement}
-   */
-  var filterForm = document.forms['upload-filter'];
-
-  /**
-   * @type {HTMLImageElement}
-   */
-  var filterImage = filterForm.querySelector('.filter-image-preview');
-
-  /**
-   * @type {HTMLElement}
-   */
-  var uploadMessage = document.querySelector('.upload-message');
-
-  /**
-   * Объявить поля ввода формы.
-   */
-  var resizeInputX = resizeForm.elements.x;
-  var resizeInputY = resizeForm.elements.y;
-  var resizeInputSize = resizeForm.elements.size;
-
-  resizeInputX.oninput = toggleResizeFormSubmit;
-  resizeInputY.oninput = toggleResizeFormSubmit;
-  resizeInputSize.oninput = toggleResizeFormSubmit;
-
-  /**
-   * @param {Action} action
-   * @param {string=} message
-   * @return {Element}
-   */
-  var showMessage = function(action, message) {
-    var isError = false;
-
-    switch (action) {
-      case Action.UPLOADING:
-        message = message || 'Кексограмим&hellip;';
-        break;
-
-      case Action.ERROR:
-        isError = true;
-        message = message || 'Неподдерживаемый формат файла<br> <a href="' + document.location + '">Попробовать еще раз</a>.';
-        break;
-    }
-
-    uploadMessage.querySelector('.upload-message-container').innerHTML = message;
-    uploadMessage.classList.remove('invisible');
-    uploadMessage.classList.toggle('upload-message-error', isError);
-    return uploadMessage;
-  };
-
-  var hideMessage = function() {
-    uploadMessage.classList.add('invisible');
-  };
-
-  /**
-   * Обработчик изменения изображения в форме загрузки. Если загруженный
-   * файл является изображением, считывается исходник картинки, создается
-   * Resizer с загруженной картинкой, добавляется в форму кадрирования
-   * и показывается форма кадрирования.
-   * @param {Event} evt
-   */
-  uploadForm.onchange = function(evt) {
-    var element = evt.target;
-    if (element.id === 'upload-file') {
-      // Проверка типа загружаемого файла, тип должен быть изображением
-      // одного из форматов: JPEG, PNG, GIF или SVG.
-      if (fileRegExp.test(element.files[0].type)) {
-        var fileReader = new FileReader();
-
-        showMessage(Action.UPLOADING);
-
-        fileReader.onload = function() {
-          cleanupResizer();
-
-          currentResizer = new Resizer(fileReader.result);
-          currentResizer.setElement(resizeForm);
-          uploadMessage.classList.add('invisible');
-
-          uploadForm.classList.add('invisible');
-          resizeForm.classList.remove('invisible');
-
-          hideMessage();
-        };
-
-        fileReader.readAsDataURL(element.files[0]);
-      } else {
-        // Показ сообщения об ошибке, если формат загружаемого файла не поддерживается
-        showMessage(Action.ERROR);
-      }
-    }
-  };
-
-  /**
-   * Обработка сброса формы кадрирования. Возвращает в начальное состояние
-   * и обновляет фон.
-   * @param {Event} evt
-   */
-  resizeForm.onreset = function(evt) {
-    evt.preventDefault();
-
-    cleanupResizer();
-    updateBackground();
-
-    resizeForm.classList.add('invisible');
-    uploadForm.classList.remove('invisible');
-  };
-
-  /**
-   * Обработка отправки формы кадрирования. Если форма валидна, экспортирует
-   * кропнутое изображение в форму добавления фильтра и показывает ее.
-   * @param {Event} evt
-   */
-  resizeForm.onsubmit = function(evt) {
-    evt.preventDefault();
-
-    if (resizeFormIsValid()) {
-      var image = currentResizer.exportImage().src;
-
-      var thumbnails = filterForm.querySelectorAll('.upload-filter-preview');
-      for (var i = 0; i < thumbnails.length; i++) {
-        thumbnails[i].style.backgroundImage = 'url(' + image + ')';
-      }
-
-      filterImage.src = image;
-
-      resizeForm.classList.add('invisible');
-      filterForm.classList.remove('invisible');
-    }
-  };
-
-  /**
-   * Сброс формы фильтра. Показывает форму кадрирования.
-   * @param {Event} evt
-   */
-  filterForm.onreset = function(evt) {
-    evt.preventDefault();
-
-    var selectedFilter = Cookies.get('upload-filter');
-    if (selectedFilter) {
-      filterImage.className = 'filter-image-preview ' + filterMap[selectedFilter];
-    }
-
-    filterForm.classList.add('invisible');
-    resizeForm.classList.remove('invisible');
-  };
-
-  /**
-   * Отправка формы фильтра. Возвращает в начальное состояние, предварительно
-   * записав сохраненный фильтр в cookie.
-   * @param {Event} evt
-   */
-  filterForm.onsubmit = function(evt) {
-    evt.preventDefault();
-
-    var selectedFilter = getSelectedFilter();
-    Cookies.set('upload-filter', selectedFilter, { expires: getDaysSinceGraceHoppersBirthday() });
-
-    cleanupResizer();
-    updateBackground();
-
-    filterForm.classList.add('invisible');
-    uploadForm.classList.remove('invisible');
-  };
-
-  function getDaysSinceGraceHoppersBirthday() {
-    var currentYear = new Date().getFullYear();
-    var birthdayThisYear = new Date(currentYear, 11, 9);
-    var today = new Date();
-    today.setHours(0, 0, 0, 0);
-    var birthday;
-    if (birthdayThisYear < today) {
-      birthday = birthdayThisYear;
-    } else {
-      birthday = birthdayThisYear.setYear(currentYear - 1);
-    }
-    return Math.floor((today - birthday) / 86400000) + 1;
-  }
-
-  /**
-   * Обработчик изменения фильтра. Добавляет класс из filterMap соответствующий
-   * выбранному значению в форме.
-   */
-  filterForm.onchange = function() {
-    if (!filterMap) {
-      // Ленивая инициализация. Объект не создается до тех пор, пока
-      // не понадобится прочитать его в первый раз, а после этого запоминается
-      // навсегда.
-      filterMap = {
-        'none': 'filter-none',
-        'chrome': 'filter-chrome',
-        'sepia': 'filter-sepia',
-        'marvin': 'filter-marvin'
+        hideMessage();
       };
+
+      fileReader.readAsDataURL(element.files[0]);
+    } else {
+      // Показ сообщения об ошибке, если формат загружаемого файла не поддерживается
+      showMessage(Action.ERROR);
     }
-
-    var selectedFilter = getSelectedFilter();
-
-    // Класс перезаписывается, а не обновляется через classList потому что нужно
-    // убрать предыдущий примененный класс. Для этого нужно или запоминать его
-    // состояние или просто перезаписывать.
-    filterImage.className = 'filter-image-preview ' + filterMap[selectedFilter];
-  };
-
-  function getSelectedFilter() {
-    var selectedFilter = [].filter.call(filterForm['upload-filter'], function(item) {
-      return item.checked;
-    })[0].value;
-    return selectedFilter;
   }
+};
+
+/**
+ * Обработка сброса формы кадрирования. Возвращает в начальное состояние
+ * и обновляет фон.
+ * @param {Event} evt
+ */
+resizeForm.onreset = function(evt) {
+  evt.preventDefault();
 
   cleanupResizer();
   updateBackground();
-})();
+
+  resizeForm.classList.add('invisible');
+  uploadForm.classList.remove('invisible');
+};
+
+/**
+ * Обработка отправки формы кадрирования. Если форма валидна, экспортирует
+ * кропнутое изображение в форму добавления фильтра и показывает ее.
+ * @param {Event} evt
+ */
+resizeForm.onsubmit = function(evt) {
+  evt.preventDefault();
+
+  if (resizeFormIsValid()) {
+    var image = currentResizer.exportImage().src;
+
+    var thumbnails = filterForm.querySelectorAll('.upload-filter-preview');
+    for (var i = 0; i < thumbnails.length; i++) {
+      thumbnails[i].style.backgroundImage = 'url(' + image + ')';
+    }
+
+    filterImage.src = image;
+
+    resizeForm.classList.add('invisible');
+    filterForm.classList.remove('invisible');
+  }
+};
+
+/**
+ * Сброс формы фильтра. Показывает форму кадрирования.
+ * @param {Event} evt
+ */
+filterForm.onreset = function(evt) {
+  evt.preventDefault();
+
+  var selectedFilter = Cookies.get('upload-filter');
+  if (selectedFilter) {
+    filterImage.className = 'filter-image-preview ' + filterMap[selectedFilter];
+  }
+
+  filterForm.classList.add('invisible');
+  resizeForm.classList.remove('invisible');
+};
+
+/**
+ * Отправка формы фильтра. Возвращает в начальное состояние, предварительно
+ * записав сохраненный фильтр в cookie.
+ * @param {Event} evt
+ */
+filterForm.onsubmit = function(evt) {
+  evt.preventDefault();
+
+  var selectedFilter = getSelectedFilter();
+  Cookies.set('upload-filter', selectedFilter, { expires: getDaysSinceGraceHoppersBirthday() });
+
+  cleanupResizer();
+  updateBackground();
+
+  filterForm.classList.add('invisible');
+  uploadForm.classList.remove('invisible');
+};
+
+function getDaysSinceGraceHoppersBirthday() {
+  var currentYear = new Date().getFullYear();
+  var birthdayThisYear = new Date(currentYear, 11, 9);
+  var today = new Date();
+  today.setHours(0, 0, 0, 0);
+  var birthday;
+  if (birthdayThisYear < today) {
+    birthday = birthdayThisYear;
+  } else {
+    birthday = birthdayThisYear.setYear(currentYear - 1);
+  }
+  return Math.floor((today - birthday) / 86400000) + 1;
+}
+
+/**
+ * Обработчик изменения фильтра. Добавляет класс из filterMap соответствующий
+ * выбранному значению в форме.
+ */
+filterForm.onchange = function() {
+  if (!filterMap) {
+    // Ленивая инициализация. Объект не создается до тех пор, пока
+    // не понадобится прочитать его в первый раз, а после этого запоминается
+    // навсегда.
+    filterMap = {
+      'none': 'filter-none',
+      'chrome': 'filter-chrome',
+      'sepia': 'filter-sepia',
+      'marvin': 'filter-marvin'
+    };
+  }
+
+  var selectedFilter = getSelectedFilter();
+
+  // Класс перезаписывается, а не обновляется через classList потому что нужно
+  // убрать предыдущий примененный класс. Для этого нужно или запоминать его
+  // состояние или просто перезаписывать.
+  filterImage.className = 'filter-image-preview ' + filterMap[selectedFilter];
+};
+
+function getSelectedFilter() {
+  var selectedFilter = [].filter.call(filterForm['upload-filter'], function(item) {
+    return item.checked;
+  })[0].value;
+  return selectedFilter;
+}
+
+cleanupResizer();
+updateBackground();
